+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - turn restriction support for restrictions with overlapping and/or multiple via-edges/ways, #3030
 - constructor of BaseGraph.Builder uses byte instead of integer count.
 - KeyValue is now KValue as it holds the value only. Note, the two parameter constructor uses one value for the forward and one for the backward direction (and no longer "key, value")
+- prioritise separate sidewalks, #3042
 
 ### 9.0 [23 Apr 2024]